pyproject.toml
[tool.poetry]
name = "online-store"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "online_store"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.2"
django-cors-headers = "^4.3.1"
django-filter = "^23.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
whitenoise = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
